% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_rhat.R
\name{all_local_rhat}
\alias{all_local_rhat}
\title{Computation of \eqn{\hat{R}(x)} on a given grid}
\usage{
all_local_rhat(chains, max_nb_points = 500)
}
\arguments{
\item{chains}{an array of size \eqn{n \times m} where \eqn{n} is the length of
the chains and \eqn{m \geq 2} is the number of chains.}

\item{max_nb_points}{the maximal length of the grid in the case where the total
number of samples is larger. By default, \code{max_nb_points = 500}.}
}
\value{
A list that contains the different \eqn{\hat{R}(x)} values used for
the computation of \eqn{\hat{R}_\infty}.
}
\description{
Return the vector of \eqn{\hat{R}(x)} values that will be used to compute
the supremum \eqn{\hat{R}_\infty}.
The function use \code{grid_for_R} for the grid of values of \eqn{x}.
}
\details{
The computation of \eqn{\hat{R}_\infty} require to compute a supremum
of \eqn{\hat{R}(x)} values over the quantiles \eqn{x}. See the documentation
of \code{grid_for_R} to see how the choice of the different \eqn{x} is done.
Thus, this function combine \code{grid_for_R} for the computation of the grid
and \code{local_rhat} for the computation of \eqn{\hat{R}(x)} on a given
\eqn{x} to obtain all the values of \eqn{\hat{R}(x)} on the generated grid.
}
\examples{
library(localrhat)

N <- 500 # length of chains
M <- 4 # number of chains

# Toy example with 3 i.i.d chains uniform in [-0.5, 0.5] and 1 in [-1, 1]:
chains <- array(c(runif((M-1)*N, -1/2, 1/2), runif(N, -1,1)), c(N,M))

# plot_local_r(grid_for_R(chains), all_local_rhat(chains),
#              xlim = c(-1, 1), ylim=c(0.999,1.1), title ="Gaussian distributions")

}
\references{
TO INCLUDE
}
