% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_rhat.R
\name{multivariate_all_local_rhat}
\alias{multivariate_all_local_rhat}
\title{Computation of multivariate \eqn{\hat{R}(x)} on a given grid}
\usage{
multivariate_all_local_rhat(chains, dir = NULL, max_nb_points = 500)
}
\arguments{
\item{chains}{an array of size \eqn{n \times d \times m}, where \eqn{n} is
the length of the chains, \eqn{m \geq 2} is the number of chains and \eqn{d}
is the dimension.}

\item{dir}{a binary vector of size \eqn{d} indicating the signs in the indicator
variable. For example, \code{dir = c(0, ..., 0)} correspond to the computation
of \eqn{\hat{R}} on \eqn{I\{\theta_1^{(j)} \leq x_1, \ldots, \theta_d^{(j)} \leq x_d\}}.
This direction will be used if no arguments is given.}

\item{max_nb_points}{the maximal length of the grid in the case where the total
number of samples is larger. By default, \code{max_nb_points = 500}.}
}
\value{
A list that contains the different \eqn{\hat{R}(x)} values used for
the computation of \eqn{\hat{R}_\infty}.
}
\description{
Return the vector of \eqn{\hat{R}(x)} values that will be used to compute
the supremum \eqn{\hat{R}_\infty} in the multivariate case.
See \code{multivariate_grid_for_R} for the grid computation for \eqn{x}.
}
\details{
The computation of \eqn{\hat{R}_\infty} require to compute a supremum
of \eqn{\hat{R}(x)} values over the quantiles \eqn{x}. See the documentation
of \code{multivariate_grid_for_R} and \code{grid_for_R} to see how the choice
of the different \eqn{x} is done.
Thus, this function combine \code{multivariate_grid_for_R} for the computation
of the grid and \code{multivariate_directed_local_rhat} or
\code{multivariate_directed_local_rhat}, depending on if a direction is specified.
}
\references{
TO INCLUDE
}
