% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_rhat.R
\name{multivariate_directed_local_rhat}
\alias{multivariate_directed_local_rhat}
\title{Multivariate \eqn{\hat{R}(x)} with a specified dependence direction}
\usage{
multivariate_directed_local_rhat(x, chains, dir)
}
\arguments{
\item{x}{a vector of size \eqn{d} corresponding to the quantile used for the computation
of \eqn{\hat{R}(x)}.}

\item{chains}{an array of size \eqn{n \times d \times m}, where \eqn{n} is
the length of the chains, \eqn{m \geq 2} is the number of chains and \eqn{d}
is the dimension.}

\item{dir}{a binary vector of size \eqn{d} indicating the signs in the indicator
variable. For example, \code{dir = c(0, ..., 0)} correspond to the computation
of \eqn{\hat{R}} on \eqn{I\{\theta_1^{(j)} \leq x_1, \ldots, \theta_d^{(j)} \leq x_d\}},
and if \code{dir[i] == 1}, then the indicator with \eqn{\theta_d^{(j)} \geq x_d}
will be used.}
}
\value{
The corresponding multivariate \eqn{\hat{R}(x)} computed on the
given dependence direction.
}
\description{
Compute \eqn{\hat{R}(x)} in the multivariate case, for a given quantile x and
a given sense for the \eqn{d} signs in the indicator variable.
}
\details{
In the multivariate case, \eqn{\hat{R}} can be computed on
\eqn{I\{\theta_1^{(j)} \leq x_1, \ldots, \theta_d^{(j)} \leq x_d\}},
but also with indicator where some "\eqn{\leq}" are replaced by "\eqn{\geq}",
which leads to \eqn{2^{d-1}} possibilities and potentially as many different
results of \eqn{\hat{R}_\infty}. Here, a 0-1 vector has to be mentioned in
the function to specify which indicator to use, indicating which size on each
dimension.
}
\examples{
library("localrhat")
library("MASS") # For Multivariate Normal Distribution
rho <- 0.9
m <- 4
n <- 100
reps <- 50

# Function to generate bivariate normal chains, with one that differs by its sigma:
gen_bvnormal_chains <- function(M, N, rho){
    sig_matrix <- (1-rho) * diag(2) + matrix(rho, nrow=2, ncol=2)
    return (array(c(mvrnorm((M-1)*N, mu = rep(0, 2), Sigma = diag(2)),
                    mvrnorm(N, mu = rep(0, 2), Sigma = sig_matrix)), c(N,2,M)))
}
chains <- gen_bvnormal_chains(m, n, rho)

# Quantile to evaluate:
x <- c(0.5, 0.5)
# multivariate_local_rhat(x, chains)
# multivariate_directed_local_rhat(x, chains)

}
\references{
TO INCLUDE
}
