% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_rhat.R
\name{multivariate_grid_for_R}
\alias{multivariate_grid_for_R}
\title{Quantile values used for the computation of  multivariate \eqn{\hat{R}_\infty}}
\usage{
multivariate_grid_for_R(chains, max_nb_points = 500)
}
\arguments{
\item{chains}{an array of size \eqn{n \times d \times m}, where \eqn{n} is
the length of the chains, \eqn{m \geq 2} is the number of chains and \eqn{d}
is the dimension.}

\item{max_nb_points}{the maximal length of the grid in the case where the total
number of samples is larger. By default, \code{max_nb_points = 500}.}
}
\value{
A list that contains the different \eqn{x} used for the computation of
\eqn{\hat{R}(x)}.
}
\description{
Return the set of points that will be used to estimate the supremum of
\eqn{\hat{R}(x)} over \eqn{x \in R^d}. The function works in the
same way as \code{grid_for_R} in the univariate case, see the corresponding
documentation for more details.
}
\references{
TO INCLUDE
}
