% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_rhat.R
\name{multivariate_local_rhat}
\alias{multivariate_local_rhat}
\title{Multivariate version of the local-\eqn{\hat{R}}: \eqn{\hat{R}(x)}}
\usage{
multivariate_local_rhat(x, chains)
}
\arguments{
\item{x}{a vector of size \eqn{d} corresponding to the quantile used for the computation
of \eqn{\hat{R}(x)}.}

\item{chains}{an array of size \eqn{n \times d \times m}, where \eqn{n} is
the length of the chains, \eqn{m \geq 2} is the number of chains and \eqn{d}
is the dimension.}
}
\value{
The multivariate \eqn{\hat{R}(x)} computed on the \eqn{m} chains.
}
\description{
Compute \eqn{\hat{R}(x)} in the multivariate case, which correspond to \eqn{\hat{R}}
computed on a multivariate indicator variable for a given quantile x.
}
\details{
The function return \eqn{\hat{R}} computed on
\deqn{I\{\theta_1^{(j)} \leq x_1, \ldots, \theta_d^{(j)} \leq x_d\}},
for a given value of \eqn{x \in R^d}.
}
\examples{
library("localrhat")
library("MASS") # For Multivariate Normal Distribution
rho <- 0.9
m <- 4
n <- 100
reps <- 50

# Function to generate bivariate normal chains, with one that differs by its sigma:
gen_bvnormal_chains <- function(M, N, rho){
    sig_matrix <- (1-rho) * diag(2) + matrix(rho, nrow=2, ncol=2)
    return (array(c(mvrnorm((M-1)*N, mu = rep(0, 2), Sigma = diag(2)),
                    mvrnorm(N, mu = rep(0, 2), Sigma = sig_matrix)), c(N,2,M)))
}
chains <- gen_bvnormal_chains(m, n, rho)

# Quantile to evaluate:
x <- c(0.5, 0.5)
# multivariate_local_rhat(x, chains)

}
\references{
TO INCLUDE
}
