% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infinity_rhat.R
\name{rhat_infinity_max_directions}
\alias{rhat_infinity_max_directions}
\title{Symmetrical version of \eqn{\hat{R}_\infty} in the multivariate case.}
\usage{
rhat_infinity_max_directions(chains, max_nb_points = 500, nb_directions = NULL)
}
\arguments{
\item{chains}{an array of size \eqn{n \times d \times m}, where \eqn{n} is
the length of the chains, \eqn{m \geq 2} is the number of chains and \eqn{d}
is the dimension.}

\item{max_nb_points}{the maximal length of the grid in the case where the total
number of samples is larger. By default, \code{max_nb_points = 500}.}

\item{nb_directions}{an integer giving the number of directions to compute
\eqn{\hat{R}_\infty}: if given, it has to be between 1 and \eqn{2^{d-1}} and a
a random set of directions of this size will be used.}
}
\value{
The maximum value of all \eqn{\hat{R}_\infty} computed.
}
\description{
Compute the multivariate version of \eqn{\hat{R}_\infty} in all possible
dependence direction and return the maximum values.
}
\details{
In the multivariate case, the sensitivity of \eqn{R_\infty} strongly
depends on the sign of dependence of the variables. This function implement
the naive way to consider all direction, by computing on the \eqn{2^{d-1}}
possible one. \strong{Warning}: this method becomes extremely costly when the
dimension grow, and we recommend to avoid it as soon as \eqn{d > 5} (pending future
work on multidimensional improvements).
Alternatively, a number of directions can be specified, to avoid computational
issues and compute \eqn{R_\infty} on a subset of the possibilities: then,
random directions will be generated to diagnostic the convergence on the
corresponding dependence direction.
}
\examples{
library(localrhat)
library("MASS") # For Multivariate Normal Distribution
rho <- 0.9
m <- 4
n <- 100
reps <- 50

# Function to generate bivariate normal chains, with one that differs by its sigma:
gen_bvnormal_chains <- function(M, N, rho){
    sig_matrix <- (1-rho) * diag(2) + matrix(rho, nrow=2, ncol=2)
    return (array(c(mvrnorm((M-1)*N, mu = rep(0, 2), Sigma = diag(2)),
                    mvrnorm(N, mu = rep(0, 2), Sigma = sig_matrix)), c(N,2,M)))
}

R_values <- c()
for (e in 1:reps){
    chains <- gen_bvnormal_chains(m, n, rho)
    # R_values <- c(R_values, rhat_infinity_max_directions(chains))
}
# R_mat <- matrix(data = R_values, ncol = 1)
# colnames(R_mat) <- "R-hat-inf_max_dir"

# plot_hist(R_mat, bin_size = 0.005,
#           lim_y_axis = 15, vaxis_pos = 1.025,
#           plot_threshold = F)

}
