% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_rhat.R
\name{trad_rhat}
\alias{trad_rhat}
\title{Compute the traditional Gelman-Rubin diagnostic.}
\usage{
trad_rhat(chains)
}
\arguments{
\item{chains}{an array of size \eqn{n \times m} where \eqn{n} is the length of
the chains and \eqn{m \geq 2} is the number of chains.}
}
\value{
\eqn{\hat{R}} computed on the \eqn{m} chains.
}
\description{
Compute the traditional Gelman-Rubin diagnostic.
}
\details{
The function return the tradition Gelman-Rubin diagnostic \eqn{\hat{R}}.
}
\references{
TO INCLUDE
}
